split status package into status & policy
status was logically doing 2 things:

1. defining the "policy" e.g. when to rotate keys, define expiry etc
2. testing against, and enforcing the policy, e.g. "this key's expiry is
too long"

I realised this when I tried to point PgpKey.Generate at status'
NextExpiryTime function and got a circular import[1]

I think what we *actually* want is a separate "policy" package which
defines the rules, and have status just enforce them. Then, PgpKey can
ask the policy package for the expiry it should use.

I think the policy package will also be a perfect place to define
the ciphers, hashes, compression preferences[2]

[1]:
https://trello.com/c/KfZguoBa/182-when-we-make-a-new-key-set-its-expiry-to-the-next-rotation-time-not-60-days
[2]:
https://trello.com/c/5IfrDbky/185-a-newly-generated-key-doesnt-have-any-cipher-hash-or-compression-preferences
